+++ src/main/java/com/project/LIA/repository/BookImgRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface BookImgRepository extends JpaRepository<BookImgDomain, Long> {
-    List<BookImgDomain> findByBook(int bookId);
+    List<BookImgDomain> findByBook(long bookId);
 }
